MiniFrame.Frames

Description

An implementation of a data frame which comes with various handy operations.

Documentation

data MiniFrame

Constructors

MiniFrame

Fields

name :: !Name

header :: !Header

rows :: ![Row]

type Name = String

type Header = [Name]

type Row = [String]

type Column = [String]

Construction

fromSample :: MiniFrame

A sample miniframe

fromNull :: MiniFrame

Built an empty miniframe

fromRows :: Name -> Header -> [Row] -> MiniFrame

Build from rows

fromColumns :: Name -> Header -> [Column] -> MiniFrame

Build from columns

fromCSV :: String -> IO MiniFrame

Build from the CSV file

Retrieval

nameOf :: MiniFrame -> Name

Get the name

headerOf :: MiniFrame -> Header

Get the header

rowsOf :: MiniFrame -> [Row]

Get the rows

columnsOf :: MiniFrame -> [Column]

Get the columns

headOf :: MiniFrame -> Row

Get the first entry

tailOf :: MiniFrame -> [Row]

Get the last entry

Dimensions

rowsNum :: MiniFrame -> Int

Get the number of rows

columnsNum :: MiniFrame -> Int

Get the number of columns

entriesNum :: MiniFrame -> Int

Get the number of entries

Modification

prependRow :: Row -> MiniFrame -> MiniFrame

Add a row to the beginning

prependColumn :: Name -> Column -> MiniFrame -> MiniFrame

Add a column to the beginning

appendRow :: Row -> MiniFrame -> MiniFrame

Add a row to the end

appendColumn :: Name -> Column -> MiniFrame -> MiniFrame

Add a column to the end

insertRow :: Index -> Row -> MiniFrame -> MiniFrame

Insert a row at the given index

insertColumn :: Index -> Name -> Column -> MiniFrame -> MiniFrame

Insert a column at the given index (index starts from 0)

renameMf :: Name -> MiniFrame -> MiniFrame

Rename the miniframe

Removal

removeRowByIndex :: Index -> MiniFrame -> MiniFrame

Remove a row by index

removeColumnByName :: Name -> MiniFrame -> MiniFrame

Remove a column by name

Pretty-printing

printName :: MiniFrame -> IO ()

Print the name

printHeader :: MiniFrame -> IO ()

Print the header

printRow :: Index -> MiniFrame -> IO ()

Print the row by index

printRows :: MiniFrame -> IO ()

Print the rows

printColumn :: Name -> MiniFrame -> IO ()

Print the column by name

printColumns :: MiniFrame -> IO ()

Print the columns

printMF :: MiniFrame -> IO ()

Print the miniframe

Additional operations

rowByIndex :: Index -> MiniFrame -> Row

Get a row by index

columnByName :: Name -> MiniFrame -> Column

Get a column by name

columnByIndex :: Index -> MiniFrame -> Column

Get the column by index

Conversion

toInt :: Name -> MiniFrame -> [Int]

Convert a column of strings to a column of integers

toDecimal :: Name -> MiniFrame -> [Float]

Convert a column of strings to a column of decimals

toBigInt :: Name -> MiniFrame -> [Integer]

Convert a column of strings to a column of big integers

toBigDecimal :: Name -> MiniFrame -> [Double]

Convert a column of strings to a column of big decimals